graph.js: Node and edge style
Add style tags of "visited" or "unvisited" to the nodes as per jaal-dijkstra-student.json.
Add style tags of "selected" or "unselected" to the edges as per jaal-dijkstra-student.json.

Update node.test.js to facilitate the style tags and write 2 test cases for this.

node.test.js: create dom_node with a classList, add this to the node. Add jsdom as jest-environment to be able to create the dom_node.
package.json: add jest-environment-jsdom as devDependency, required for the node.test.js file.

Tests are passing, ran with `npm test`.