Record jsavcanvas as svg screenshots
Fulfil #22 .

`animation/svg.js`: Extract the jsav data from the current page and compile that into an svg image. This svg is returned as a string.
`animation/test_svg.svg`: Sample svg data produced.
`animation/animation.js`: Call for the svg to be produced on every gradableStep. Add this data to the data submitted to the server.