Exercise player related fixes
`metadata.js`: Correctly set the jaalVersion.
`submission.js`: add svg for the initialState.
`initialState.js`: add svg for the initialState
`model-svg.js`: limit the width of the svg by tallying the width of all of the elements
+ a 20 px margin per element.
`model-answer-definition.js`: add svg to the first step recorded (narration step).
`exerciseRecorder.js`: disable `finishWithoutModelAnswer` as it was causing
double click-handler bugs in the player.
`svg.js`: remove adding svg to end of page, was used for dev.