Implement model answer with priority queue
`v2.json`: Add new model answer strings. Finnish ones still need to
be translated.
`v2.css`: Shift white box from priority queue to the alignment
element, so as to not have the border in the model answer.
`v2.js`: Add priority queue to model answer, and rewrite model
answer to use this. Each neighbouring node that is visited
in the model answer is now explicitly explained what is done with
it. Finnish translation for this still needs to be done.
`getTreeNodeList` is moved to outside of the update button,
so that it can be accessed by the model answer.
`moveUp` is replaced with inline implementation, as passing
a node and then calling the parent function breaks the grading
on the ODSA side.
`findParent` gets a new parameter for which heap it is in,
so it can be used by model answer and the student's work.
`heapsize` has been changed into a jsav variable, so as to
not break the priority queue when steps are undone.
All uses of the variable are changed to reflect this.