Save multiline text in SVG
Adapt the way nodes are recorded to parse the node label:
replace any HTML <br> with using SVG <tspan> with `dy` offsets.

Next to that, adapt the offset of the text to be positive if
only single line, and negative if multiple lines. This way the
text in the node is still centered.