Resolve issue #98
Resolve the issue where if a node was in the prio queue several times,
it can be that the wrong edge gets added compared to the node
removed from the binary tree. the delete button now reads the
value from the root of the binary tree, rather than from the table.