Remove async from escapeRecording()
Function escapeRecording() had an "async" keyword, because Artturi thought
zlib can only work asynchronously. (RTFM!) The zipping in side that function
is now synchronous, so the function does not need to be asynchronous. This
fixed a bug in the production setting: "answer" field in the HTTP POST request
was missing, because the type of it was Promise.