Resolve issues #107 and #104
After the model answer was opened, the JAAL data would no longer
contain the model answer for that session. This was due to the
modelAnswer.opened flag being set to true the first time the
model answer was opened, and never set to false. Only with
this flag on false will the recorder record the model answer.
Now on reset, modelAnswer.opened and modelAnswer.ready are
set to false, so that the model answer will be recorded even if
the model solution has been looked at by the student for an earlier
exercise that session.

Additionally also reset the JAAL ids when the reset button is
pressed. The JAAL id counters are reset to 1, and the jsav to jaal
id mapping is cleared when reset is pressed.