Resolve #110 for Scaffolded Prim
The fix that was implemented for scaffolded Dijkstra is now applied
to scaffolded Prim. The fix should not be necessary for non-scaffolded
Dijkstra or Prim, since they do not make use of the priority queue,
and rather use the list implementation in the model answer.

I have tested with Chrome and Firefox under Ubuntu 20 and get
alphabetical processing for scaffolded Prim.