Clean up `submisison.js` with regards to JAAL1.0 validation
Remove all code related to JAAL1.0 validation methods in
`submission.js`, as the data is validated against JAAL2.0 upon
submission and this already encompasses the validation that is
done here.

This means that we can update the JAAL schemas without having
to update the validation functions.

This change also removes some functions from the file, namely
the ones related to the JAAL 1.0 recording of HTML of the
exercise.

The changes have been tested with scaffolded and unscaffolded
Prim and Dijkstra. All four exercises still produce valid JAAL2.0.