Add Johanna's spreadsheet
Spreadsheet that contains information on which exercise contains
which JSAV data structures for the grading.

A red field in the recording semantically means that the data
structure that is still in JAAL1 or it has data structures
that are not currently supported by the grader. White means that
the data structure is currently supported, and thus the recorder
should be able to record that exercise with only minor tweaking.