Dijkstra (& Prim): Update behaviour when not in PQ
Resolve issue #79
Rather than quietly not doing anything when the vertex is not in
the priority queue, the learner now gets a warning message stating
that it is not possible to update said node as it does not exist
in the priority queue.

This option does not take away agency of the student, nor their
ability to make the mistake; rather it explains them why it is
not possible to execute that action.